+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
  * Added `batchUnstakeNfts()` function to be able to perform batch unstaking of NFTs.
  * Internal hook functions `_onStake` and`_onUnstake`.
+ * Added `withdrawLostCycleRewards` for an admin to withdraw the rewards for a past cycle where there was no stake.
 
 ### Breaking changes
  * New events `NftsBatchStaked` and `NftsBatchUnstaked` emitted during batch operations instead of the single versions.